% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignCellType.R
\docType{methods}
\name{AssignCellType}
\alias{AssignCellType}
\alias{assigncelltype}
\title{AssignCellType}
\usage{
AssignCellType(markers, org = "hsa", celltype.markers = NA)
}
\arguments{
\item{markers}{A data frame which includes differential expressed genes in each cell cluster.}

\item{org}{"hsa" or "mmu".}

\item{celltype.markers}{A list, which customize marker of all different cell type.}
}
\value{
A Vector of assigned cell type names.
}
\description{
Assign Cell Type for single cells
}
\examples{

}
\author{
Wubing Zhang
}
