% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Count2TPM.R
\docType{methods}
\name{Count2TPM}
\alias{Count2TPM}
\title{Transform raw count to TPM}
\usage{
Count2TPM(countMat, idType = "Ensembl", org = "hsa")
}
\arguments{
\item{countMat}{A matrix-like object.}

\item{idType}{Ensembl, Symbol, or Entrez.}

\item{org}{One of "hsa", "mmu", "bta", "cfa", "ptr", "rno" and "ssc".}
}
\value{
A TPM matrix.
}
\description{
Transform raw count to TPM
}
\examples{

}
