% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DysfunctionView.R
\name{DysfunctionView}
\alias{DysfunctionView}
\title{Cell surface expression of TIM-3 and PD1 can be used to partition CD8+
T cells into three different groups: double negative, double positive,
and single positive.}
\usage{
DysfunctionView(expr, xcutoff, ycutoff, fill = "CYT", title = NULL,
  filename = NULL, ...)
}
\description{
Cell surface expression of TIM-3 and PD1 can be used to partition CD8+
T cells into three different groups: double negative, double positive,
and single positive.
}
