% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getImmMeta.R
\name{getImmMeta}
\alias{getImmMeta}
\title{Select cells with high CD3 expression(T cell marker), non-zero expression
of CYT, and unequal expression of CD8 and CD4}
\usage{
getImmMeta(mat, filter = TRUE)
}
\description{
Select cells with high CD3 expression(T cell marker), non-zero expression
of CYT, and unequal expression of CD8 and CD4
}
