% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuratPipeline.R
\docType{methods}
\name{SeuratPipeline}
\alias{SeuratPipeline}
\alias{seuratpipeline}
\title{Create Seurat object}
\usage{
SeuratPipeline(tpmMat, proj, min.c = 3, min.g = 500, max.g = 20000,
  normalization.method = NULL, do.scale = TRUE, do.center = FALSE,
  vars.to.regress = c("nUMI"), outdir = ".", ...)
}
\arguments{
\item{tpmMat}{Matrix like data object, or a file path of data.}

\item{proj}{Project name (string).}

\item{min.c}{Include genes with detected expression in at least this many cells.
Will subset the raw.data matrix as well. To reintroduce excluded genes, create a
new object with a lower cutoff.}

\item{min.g}{Include cells where at least this many genes are detected.}

\item{normalization.method}{Method for cell normalization. Default is no normalization.
In this case, run NormalizeData later in the workflow. As a shortcut, you can specify
a normalization method (i.e. LogNormalize) here directly.}

\item{do.scale}{In object@scale.data, perform row-scaling (gene-based z-score). FALSE
by default. In this case, run ScaleData later in the workflow. As a shortcut, you can
specify do.scale = TRUE (and do.center = TRUE) here.}

\item{do.center}{In object@scale.data, perform row-centering (gene-based centering).}

\item{vars.to.regress}{Variables to be regressed.}

\item{outdir}{The output directory.}

\item{...}{Other useful paramters in method `CreateSeuratObject`.}
}
\value{
A Seurat object.
}
\description{
Run Seurat pipeline and reture Seurat object
}
\examples{

}
\author{
Wubing Zhang
}
