% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tSNE.R
\docType{methods}
\name{tSNE}
\alias{tSNE}
\alias{tsne}
\title{tSNE}
\usage{
tSNE(SeuratObj, dims.use = 1:10, org = "hsa", outdir = ".")
}
\arguments{
\item{SeuratObj}{A Seurat object.}

\item{dims.use}{Which dimensions to use as input features.}

\item{org}{"hsa" or "mmu".}

\item{outdir}{The output directory.}
}
\value{
A list contrains two objects, including \code{SeuratObj} and \code{markers}.
}
\description{
tSNE plot
}
\examples{

}
\author{
Wubing Zhang
}
